Major Updates

+++ Helpers/Parser.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-import string
+import os
 import re
 import subprocess
 
@@ -24,17 +24,26 @@ class Parser:
         return Result
 
     def GrepFindEmails(self):
+        # Major hack during testing;
+        # I found grep is was better at Regex than re in python
+        FinalOutput = []
+        val = ""
         with open("temp-Output.temp", "wr") as myfile:
             myfile.write(self.InputData)
         ps = subprocess.Popen(
-            ('grep', '-r', "@", "temp-Output.temp"), stdout=subprocess.PIPE)
-        val = subprocess.check_output(("grep", "-i", "-o", '[A-Z0-9._%+-]\+@[A-Z0-9.-]\+\.[A-Z]\{2,4\}'),
-                                      stdin=ps.stdout)
-        with open('temp.txt', "wr+") as myfile:
-            myfile.write(str(val))
-        with open('temp.txt', "r") as myfile:
-            output = myfile.readlines()
-        os.remove('temp.txt')
-        for item in output:
-            FinalOutput.append(item.rstrip("\n"))
+            ('grep', "@", "temp-Output.temp"), stdout=subprocess.PIPE)
+        try:
+            val = subprocess.check_output(("grep", "-i", "-o", '[A-Z0-9._%+-]\+@[A-Z0-9.-]\+\.[A-Z]\{2,4\}'),
+                                          stdin=ps.stdout)
+        except:
+            pass
+        os.remove('temp-Output.temp')
+        if len(val) > 0:
+            with open('temp.txt', "wr+") as myfile:
+                myfile.write(str(val))
+            with open('temp.txt', "wr+") as myfile:
+                output = myfile.readlines()
+            os.remove('temp.txt')
+            for item in output:
+                FinalOutput.append(item.rstrip("\n"))
         return FinalOutput
